more tests for match logic.
